+++ src/CalculatorApp.form
@@ -3,7 +3,7 @@
   <grid id="27dc6" binding="mainJPanel" layout-manager="GridLayoutManager" row-count="3" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="40" left="40" bottom="40" right="40"/>
     <constraints>
-      <xy x="20" y="20" width="771" height="400"/>
+      <xy x="20" y="20" width="771" height="481"/>
     </constraints>
     <properties/>
     <border type="none"/>
@@ -49,6 +49,15 @@
                 </constraints>
                 <properties/>
               </component>
+              <component id="33453" class="javax.swing.JLabel" binding="imageLabel2">
+                <constraints>
+                  <grid row="0" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
+                </constraints>
+                <properties>
+                  <icon value="images/taco.jpg"/>
+                  <text value=""/>
+                </properties>
+              </component>
             </children>
           </grid>
           <grid id="ebb60" binding="item3Panel" layout-manager="GridLayoutManager" row-count="2" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
@@ -79,6 +88,15 @@
                 </constraints>
                 <properties/>
               </component>
+              <component id="449ca" class="javax.swing.JLabel" binding="imageLabel3">
+                <constraints>
+                  <grid row="0" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
+                </constraints>
+                <properties>
+                  <icon value="images/hamburger.jpg"/>
+                  <text value=""/>
+                </properties>
+              </component>
             </children>
           </grid>
           <vspacer id="3e3a3">
@@ -108,11 +126,20 @@
                   <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="6" hsize-policy="1" anchor="0" fill="2" indent="0" use-parent-layout="false"/>
                 </constraints>
               </vspacer>
-              <component id="71cda" class="javax.swing.JSpinner" default-binding="true">
+              <component id="71cda" class="javax.swing.JSpinner" binding="spinner1">
                 <constraints>
                   <grid row="1" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="8" fill="1" indent="0" use-parent-layout="false"/>
                 </constraints>
                 <properties/>
+              </component>
+              <component id="e5da8" class="javax.swing.JLabel" binding="imageLabel1">
+                <constraints>
+                  <grid row="0" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
+                </constraints>
+                <properties>
+                  <icon value="images/sandwich.jpg"/>
+                  <text value=""/>
+                </properties>
               </component>
             </children>
           </grid>
