+++ src/CalculatorApp.java
@@ -1,6 +1,9 @@
 import javax.swing.*;
+import java.awt.*;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
 
-public class CalculatorApp {
+public class CalculatorApp extends JFrame implements ActionListener, Runnable {
     public JPanel mainJPanel;
     private JSpinner spinner2;
     private JSpinner spinner3;
@@ -19,4 +22,40 @@ public class CalculatorApp {
     private JLabel costLabel;
     private JLabel titleLabel;
     private JPanel titleJPanel;
+    private JSpinner spinner1;
+    private JLabel imageLabel1;
+    private JLabel imageLabel3;
+    private JLabel imageLabel2;
+
+    // JFrame 이용
+
+    private void start(){
+        // x 클릭 시 프로그램도 종료, Frame의 Layout : BorderLayout
+        JFrame frame = new JFrame("App");
+        frame.setContentPane(new CalculatorApp().mainJPanel);
+        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        frame.pack();
+
+        frame.setBackground(Color.LIGHT_GRAY);
+        frame.setSize(800, 800);
+
+        frame.setVisible(true);
+    }
+
+    public static void main(String[] args){
+        CalculatorApp client = new CalculatorApp();
+        client.start();
+    }
+
+    @Override
+    public void actionPerformed(ActionEvent e) {
+
+    }
+
+    @Override
+    public void run() {
+
+    }
 }
+
+
